move editor container into its own component
